Update middleware to add real_user attr to requests
If you need to get the real user making the request (e.g. for logging),
you can end up with a lot of code that looks like this:

    real_user = getattr(request, 'impersonator', request.user)

This commit adds the real_user to the request so that it's readily available. If
the request is being impersonated then real_user will return the impersonator,
else it will return the request.user.